Sync to master

+++ files/scripts/syncd.sh
@@ -55,15 +55,16 @@ function wait_for_database_service()
 
 function getBootType()
 {
-    case "$(cat /proc/cmdline | grep -o 'SONIC_BOOT_TYPE=\S*' | cut -d'=' -f2)" in
-    warm*)
+    # same code snippet in files/build_templates/docker_image_ctl.j2
+    case "$(cat /proc/cmdline)" in
+    *SONIC_BOOT_TYPE=warm*)
         TYPE='warm'
         ;;
-    fastfast)
+    *SONIC_BOOT_TYPE=fastfast*)
         TYPE='fastfast'
         ;;
-    fast*)
-        TYPE='fast'
+    *SONIC_BOOT_TYPE=fast*|*fast-reboot*)
+        TYPE=$(awk '{ if ($1 <= 180) print "fast"; else print "cold" }' /proc/uptime)
         ;;
     *)
         TYPE='cold'
@@ -101,8 +102,13 @@ start() {
         fi
 
         if [[ x"$WARM_BOOT" != x"true" ]]; then
-            /bin/systemctl stop pmon
-            /usr/bin/hw-management.sh chipdown
+            if [[ x"$(/bin/systemctl is-active pmon)" == x"active" ]]; then
+                /bin/systemctl stop pmon
+                /usr/bin/hw-management.sh chipdown
+                /bin/systemctl restart pmon
+            else
+                /usr/bin/hw-management.sh chipdown
+            fi
         fi
 
         if [[ x"$BOOT_TYPE" == x"fast" ]]; then
@@ -112,10 +118,6 @@ start() {
         /usr/bin/mst start
         /usr/bin/mlnx-fw-upgrade.sh
         /etc/init.d/sxdkernel start
-
-        if [[ x"$WARM_BOOT" != x"true" ]]; then
-            /bin/systemctl start pmon
-        fi
     fi
 
     if [[ x"$WARM_BOOT" != x"true" ]]; then
@@ -127,10 +129,6 @@ start() {
     # start service docker
     /usr/bin/${SERVICE}.sh start
     debug "Started ${SERVICE} service..."
-
-    if [[ x"$sonic_asic_platform" == x"mellanox" && x"$BOOT_TYPE" == x"fast" ]]; then
-        /usr/bin/hw-management.sh chipupen
-    fi
 
     unlock_service_state_change
 }
